fix some sles11sp4 issues
